+++ packages/vda-common/src/contract.ts
@@ -63,9 +63,9 @@ export const CONTRACT_ADDRESS : Record<CONTRACT_NAMES, Record<string, string | n
     "StorageNodeRegistry": {
         mainnet: null,
         "0x89": null,
-        testnet: "0xDfe326357A4078A3ea8A192084255F25Ab73c712",
-        "0x13881": "0xDfe326357A4078A3ea8A192084255F25Ab73c712",
-        devnet: "0xDfe326357A4078A3ea8A192084255F25Ab73c712"
+        testnet: "0x4Fa02CA7fD115b4cCA7F80Cb3047550648c360e1",
+        "0x13881": "0x4Fa02CA7fD115b4cCA7F80Cb3047550648c360e1",
+        devnet: "0x4Fa02CA7fD115b4cCA7F80Cb3047550648c360e1"
     }
 };