+++ packages/vda-common-test/CHANGELOG.md
@@ -1,6 +1,9 @@
 2024-03-15 (v3.0.1)
 -------------------
+- Added variables for locking test in the `StorageNodeRegistry` contract
 
+2024-03-15 (v3.0.1)
+-------------------
 - Update variable names for `StorageNode`
 - Update constants for `StorageNode`
 - Update type of `nonce` from `number` to `BigNumberish`